Add scaffolding for custom routes mixin. Remove url class. Add url mixin. Move crud mixin to folder
