+++ src/core/mixins/custom-routes.js
@@ -1,0 +1,5 @@
+// @ts-check
+
+export function CustomRoutesMixin(Rapid) {
+
+}
